requirements.txt
amqp==5.0.1
asgiref==3.2.10
autopep8==1.5.4
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.43.0
django-celery-beat==2.1.0
django-silk==4.1.0
django-timezone-field==4.0
djangorestframework==3.12.1
gprof2dot==2019.11.30
idna==2.10
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
newsapi-python==0.2.6
oauthlib==3.1.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
